+++ src/midi-qol.ts
@@ -19,7 +19,7 @@ import { initHooks, readyHooks } from './module/Hooks.js';
 import { initGMActionSetup, setupSocket, socketlibSocket } from './module/GMAction.js';
 import { setupSheetQol } from './module/sheetQOL.js';
 import { TrapWorkflow, DamageOnlyWorkflow, Workflow } from './module/workflow.js';
-import { applyTokenDamage, checkNearby, findNearby, getDistance, getTraitMult, MQfromActorUuid, MQfromUuid } from './module/utils.js';
+import { applyTokenDamage, checkNearby, findNearby, getDistance, getDistanceSimple, getTraitMult, MQfromActorUuid, MQfromUuid } from './module/utils.js';
 import { ConfigPanel } from './module/apps/ConfigPanel.js';
 import { showItemCard, showItemInfo } from './module/itemhandling.js';
 import { RollStats } from './module/RollStats.js';
@@ -192,7 +192,7 @@ function setupMidiQOLApi() {
     configSettings: () => {return configSettings},
     ConfigPanel: ConfigPanel,
     getTraitMult: getTraitMult,
-    getDistance: getDistance,
+    getDistance: getDistanceSimple,
     midiFlags,
     debug,
     log,